YoloV8WebProject/requirement.txt
aioice==0.9.0
aiortc==1.5.0
altair==4.2.2
attrs==23.1.0
av==10.0.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
filelock==3.12.0
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
google-crc32c==1.5.0
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lit==16.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
protobuf==3.20.3
psutil==5.9.5
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.1b0
pyee==9.0.4
Pygments==2.15.1
pylibsrtp==0.8.0
Pympler==1.0.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.29.0
rich==13.3.5
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.21.1
six==1.16.0
smmap==5.0.0
streamlit==1.21.0
streamlit-webrtc==0.45.0
sympy==1.11.1
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
triton==2.0.0
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ultralytics==8.0.91
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
